feature(labmda): Respect event.isBase64Encoded flag
AWS Lambdas have the option for the body being sent to the Function to be encoded in Base64.  I'm trying to use this feature with Zeit's `now dev` and they are, indeed, Base64 encoding the body.  Funnily enough, this is not the case when your run your lambda function within their production environment.  Because of AWS having this flag, I thought it was appropriate to make a PR here rather than have Zeit change their practices.